refactor: sortCasesByAge as standalone sorting function
